:wrench: skill Xp needed

+++ src/Script/Class/SubClass/Skill.cs
@@ -10,6 +10,6 @@ namespace SpaceColony.Script.Class.SubClass
     {
         public int Point { get; set; }
         public int Interrest  { get; set; }
-        public double Xp { get; set; }
+        public int Xp { get; set; }
     }
 }
